Removed Python 3.3-specific testing.
I believe the way things were setup before my deactivating and re-activating the virtual environment lead to it using Python 2.7 even when it said it was going to use Python 3.3. For now, I can't come up with a good way to test on Python 3 without recompiling numpy each time, which is insanely time consuming.