Bumped up version number in setup.py to make pip install latest version with Python 3 fixes.
Currently, the Python ml_metrics package on PyPI is not Python 3 compatible because it is out-of-date. I've bumped the version number up to avoid any conflicts, so if you could please run "python setup.py register" and "python setup.py sdist upload" so this latest version gets up there, that would be extremely helpful.

We at ETS have recently released a ML package, SciKit-Learn Laboratory (SKLL) that relies on ml_metrics for kappa, and we don't want to have to repackage your code with ours unless absolutely necessary.